All Java files for the Doughnut Game.
The game is ready to play. The dist folder is included, so just launch the app from there. All files are open-source, so feel free to use them as you like, but please, I would really appreciate if you could mention me as the creator and this repository as the link, since that would help me lots to launch my career as a programmer. Thanks!